Update app.py
There is a really weird mistake on this archive but it won't be like that for too long. Now all the functions for user maintenance, csv imports and exports and even small functions for machines and groups can all be found in a single archive. It is lacking functions but it will be fixed pretty soon.